integer distance (for now)

+++ project/data_input.py
@@ -88,7 +88,7 @@ def calculate_euc_distance(a, b):
     x1, y1 = a
     x2, y2 = b
 
-    return math.sqrt(((x1 - x2) ** 2) + (((y1 - y2) ** 2)))
+    return int(math.sqrt(((x1 - x2) ** 2) + (((y1 - y2) ** 2))))
 
 
 def _post_process_specs(specs):
